+++ src/control_plane/daemon/unix_daemon_runtime.rs
@@ -3,12 +3,12 @@ use super::{
     BollardUnixEngineConnector, DaemonIterationResult, DaemonRequestDispatchOptions,
     DiscoveryScheduler, EngineConnectionOutcome, EngineConnectionSupervisor,
     EngineImageReferenceResolution, EngineReconciliationPlanOptions, EngineReconciliationSchedule,
-    FilesystemEventWatcher, ImageReferenceResolution, IpcEventJournal, ProjectBackupQueue,
-    ProjectCommandQueue, ProjectLogSessionRegistry, ProjectRestoreQueue, ResourceHealthRegistry,
-    RetryBackoff, RetryBackoffOptions, SingletonLease, UnixDaemonRuntimeError,
-    UnixDaemonRuntimeOptions, dispatch_daemon_request, initialize_default_installation,
-    invalidate_engine_connection, plan_engine_reconciliation, reconcile_watched_roots,
-    requires_followup_reconciliation, restore_daemon_operation_queues,
+    FilesystemEventWatcher, ImageReferenceResolution, IpcEventJournal, MigrationDecisionQueue,
+    ProjectBackupQueue, ProjectCommandQueue, ProjectLogSessionRegistry, ProjectRestoreQueue,
+    ResourceHealthRegistry, RetryBackoff, RetryBackoffOptions, SingletonLease,
+    UnixDaemonRuntimeError, UnixDaemonRuntimeOptions, dispatch_daemon_request,
+    initialize_default_installation, invalidate_engine_connection, plan_engine_reconciliation,
+    reconcile_watched_roots, requires_followup_reconciliation, restore_daemon_operation_queues,
     validate_project_workload_adoption,
 };
 use crate::control_plane::application::ControlPlane;
@@ -57,6 +57,7 @@ pub(crate) struct UnixDaemonRuntime {
     pub(super) project_commands: ProjectCommandQueue,
     pub(super) project_backups: ProjectBackupQueue,
     pub(super) project_restores: ProjectRestoreQueue,
+    pub(super) migration_decisions: MigrationDecisionQueue,
     pub(super) project_logs: ProjectLogSessionRegistry,
     pub(super) resource_health: ResourceHealthRegistry,
     pub(super) active_project_logs: BTreeMap<String, ActiveProjectLogSession>,
@@ -90,7 +91,7 @@ impl UnixDaemonRuntime {
             &runtime_directory.join("state-backups"),
             unix_time_seconds(),
         )?;
-        let (project_commands, project_backups, project_restores) =
+        let (project_commands, project_backups, project_restores, migration_decisions) =
             restore_daemon_operation_queues(&mut store, unix_time_seconds())?;
         let event_journal = IpcEventJournal::restore(store.daemon_events()?)?;
         let installation = initialize_default_installation(&mut store)?;
@@ -124,6 +125,7 @@ impl UnixDaemonRuntime {
             project_commands,
             project_backups,
             project_restores,
+            migration_decisions,
             project_logs: ProjectLogSessionRegistry::default(),
             resource_health: ResourceHealthRegistry::default(),
             active_project_logs: BTreeMap::new(),
@@ -173,6 +175,7 @@ impl UnixDaemonRuntime {
                 project_commands: &mut self.project_commands,
                 project_backups: &mut self.project_backups,
                 project_restores: &mut self.project_restores,
+                migration_decisions: &mut self.migration_decisions,
                 project_logs: &mut self.project_logs,
                 resource_health: &self.resource_health,
                 image_reference_resolution: image_reference_resolution.as_mut().map(|resolver| {